backend/requirements.txt
flask[async]
flask
flask-pymongo
python-dotenv 
flask-limiter
google-generativeai